refactor(meta_store): Reduce the dependency of meta_store
